unicode smileys never worked in MyDetails
git-svn-id: http://svn.miranda-ng.org/main/trunk@3382 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/MyDetails/src/data.cpp
@@ -105,10 +105,9 @@ Protocol::~Protocol()
 
 void Protocol::lcopystr(TCHAR *dest, TCHAR *src, size_t maxlen)
 {
-	if (lstrcmp(dest, src) != 0)
-	{
+	if (lstrcmp(dest, src) != 0) {
 		data_changed = true;
-		lstrcpyn(dest, src, maxlen);
+		lstrcpyn(dest, src, (DWORD)maxlen);
 	}
 }
